seguridad/requirements.txt
Flask==3.0.2
PyJWT==2.8.0
cryptography==42.0.5
prometheus-client==0.20.0